+++ examples/standalone_live.py
@@ -1,0 +1,17 @@
+from time import sleep
+from turtle import done, forward, left, right
+
+
+forward(100)
+sleep(1)
+
+left(90)
+sleep(1)
+
+forward(100)
+sleep(1)
+
+right(45)
+forward(70)
+
+done()